auth_server/requirements.txt
fastapi
prometheus-fastapi-instrumentator
uvicorn
sqlmodel
mysql-connector-python
cryptography
PyJWT
passlib[bcrypt]
redis
